// src/types/VendorTypes.ts
export interface Contact {
  name: string;
  phone: string;
  email: string;
  fax: string;
  designation: string;
}

export interface Vendor {
  id: number;
  type: string;
  legal_name: string;
  remit_name: string;
  vendor_type: string;
  service: string;
  primary_address: string;
  primary_city: string;
  primary_state: string;
  primary_country: string;
  primary_postal: string;
  primary_email: string;
  primary_phone: string;
  primary_fax: string;
  scac: string;
  docket_number: string;
  vendor_code: string;
  gst_hst_number: string;
  qst_number: string;
  ca_bond_number: string;
  website: string;
  sameAsPrimary: boolean;
  mailing_address: string;
  mailing_city: string;
  mailing_state: string;
  mailing_country: string;
  mailing_postal: string;
  mailing_email: string;
  mailing_phone: string;
  mailing_fax: string;
  us_tax_id: string;
  payroll_no: string;
  wcb_no: string;
  ar_name: string;
  ar_email: string;
  ar_contact_no: string;
  ar_ext: string;
  ap_name: string;
  ap_email: string;
  ap_contact_no: string;
  ap_ext: string;
  bank_name: string;
  bank_phone: string;
  bank_email: string;
  bank_us_acc_no: string;
  bank_cdn_acc_no: string;
  bank_address: string;
  cargo_company: string;
  cargo_policy_start: string;
  cargo_policy_end: string;
  cargo_ins_amt: number;
  liab_company: string;
  liab_policy_start: string;
  liab_policy_end: string;
  liab_ins_amt: number;
  contacts: Contact[];
  created_at: string;
  updated_at: string;
}